requirements.txt
streamlit>=1.54.0,<2.0
pandas>=2.2.0,<3.0
python-dotenv>=1.0.0,<2.0
mistralai>=1.2.0,<2.0
pytest>=8.0.0,<9.0
